+++ src/contexts/index.ts
@@ -2,6 +2,7 @@ export { EdgesContext, useEdgesProperties } from './EdgesContext';
 export { GraphMetaContext, useGraphMeta } from './GraphMetaContext';
 export { GraphRegistryContext, useGraphRegistry } from './GraphRegistryContext';
 export { GraphSelectionContext, useGraphSelection } from './GraphSelectionContext';
+export { GraphWorkspaceContext, useGraphWorkspace } from './GraphWorkspaceContext';
 export { LayoutContext, useLayoutProperties } from './LayoutContext';
 export { ModalsContext, useModals } from './ModalsContext';
 export { NodesContext, useNodeProperties } from './NodesContext';