+++ src/providers/PropertiesProvider.tsx
@@ -3,22 +3,25 @@ import { EdgesProvider } from './EdgesProvider';
 import { GraphMetaProvider } from './GraphMetaProvider';
 import { GraphRegistryProvider } from './GraphRegistryProvider';
 import { GraphSelectionProvider } from './GraphSelectionProvider';
+import { GraphWorkspaceProvider } from './GraphWorkspaceProvider';
 import { LayoutProvider } from './LayoutProvider';
 import { NodesProvider } from './NodesProvider';
 
 export function PropertiesProvider({ children }: PropertiesProviderProps) {
     return (
-        <GraphRegistryProvider>
-            <GraphMetaProvider>
-                <GraphSelectionProvider>
-                    <LayoutProvider>
-                        <NodesProvider>
-                            <EdgesProvider>{children}</EdgesProvider>
-                        </NodesProvider>
-                    </LayoutProvider>
-                </GraphSelectionProvider>
-            </GraphMetaProvider>
-        </GraphRegistryProvider>
+        <GraphWorkspaceProvider>
+            <GraphRegistryProvider>
+                <GraphMetaProvider>
+                    <GraphSelectionProvider>
+                        <LayoutProvider>
+                            <NodesProvider>
+                                <EdgesProvider>{children}</EdgesProvider>
+                            </NodesProvider>
+                        </LayoutProvider>
+                    </GraphSelectionProvider>
+                </GraphMetaProvider>
+            </GraphRegistryProvider>
+        </GraphWorkspaceProvider>
     );
 }